requirements.txt
adbutils==1.2.15
altgraph==0.17.4
apkutils2==1.0.0
av==9.2.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cigam==0.0.3
click==8.1.8
contourpy==1.3.0
cryptography==44.0.3
cx_Freeze==8.3.0
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
dmgbuild==1.6.5
ds-store==1.3.1
ffmpeg==1.4
filelock==3.18.0
firebase-admin==6.8.0
Flask==3.1.0
fonttools==4.58.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
mac-alias==2.2.2
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.9.4
msgpack==1.1.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
passlib==1.7.4
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
py==1.11.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycrypto==2.6.1
pyelftools==0.32
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retry==0.9.2
rsa==4.9.1
scikit-image==0.24.0
scipy==1.13.1
scrcpy-client==0.4.7
SimpleCV==1.3
six==1.17.0
sv-ttk==2.6.0
tifffile==2024.8.30
tkcalendar==1.6.1
TKinterModernThemes==1.10.4
tomli==2.2.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
whichcraft==0.6.1
xmltodict==0.14.2
zipp==3.21.0
